+++ golem-worker-service-base/src/service/worker/default.rs
@@ -51,7 +51,7 @@ pub trait WorkerService<AuthCtx> {
     async fn create(
         &self,
         worker_id: &WorkerId,
-        template_version: i32,
+        template_version: u64,
         arguments: Vec<String>,
         environment_variables: HashMap<String, String>,
         metadata: WorkerRequestMetadata,
@@ -204,7 +204,7 @@ where
     async fn create(
         &self,
         worker_id: &WorkerId,
-        template_version: i32,
+        template_version: u64,
         arguments: Vec<String>,
         environment_variables: HashMap<String, String>,
         metadata: WorkerRequestMetadata,
@@ -1324,7 +1324,7 @@ where
     async fn create(
         &self,
         _worker_id: &WorkerId,
-        _template_version: i32,
+        _template_version: u64,
         _arguments: Vec<String>,
         _environment_variables: HashMap<String, String>,
         _metadata: WorkerRequestMetadata,
